:truck: (GQL) Rename typeDefs ts file
